executor: support `CreateUserStmt`, `SetPwdStmt`.
