+++ executor/converter/convert_stmt.go
@@ -850,14 +850,6 @@ func convertShow(converter *expressionConverter, v *ast.ShowStmt) (*stmts.ShowSt
 	return oldShow, nil
 }
 
-func convertSetPwd(converter *expressionConverter, v *ast.SetPwdStmt) (*stmts.SetPwdStmt, error) {
-	return &stmts.SetPwdStmt{
-		User:     v.User,
-		Password: v.Password,
-		Text:     v.Text(),
-	}, nil
-}
-
 func convertUserSpec(converter *expressionConverter, v *ast.UserSpec) (*coldef.UserSpecification, error) {
 	oldSpec := &coldef.UserSpecification{
 		User: v.User,
@@ -871,21 +863,6 @@ func convertUserSpec(converter *expressionConverter, v *ast.UserSpec) (*coldef.U
 		oldSpec.AuthOpt = oldAuthOpt
 	}
 	return oldSpec, nil
-}
-
-func convertCreateUser(converter *expressionConverter, v *ast.CreateUserStmt) (*stmts.CreateUserStmt, error) {
-	oldCreateUser := &stmts.CreateUserStmt{
-		IfNotExists: v.IfNotExists,
-		Text:        v.Text(),
-	}
-	for _, val := range v.Specs {
-		oldSpec, err := convertUserSpec(converter, val)
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		oldCreateUser.Specs = append(oldCreateUser.Specs, oldSpec)
-	}
-	return oldCreateUser, nil
 }
 
 func convertPrivElem(converter *expressionConverter, v *ast.PrivElem) (*coldef.PrivElem, error) {
