--- stmt/stmts/account_manage.go
@@ -1,166 +1,0 @@
-// Copyright 2013 The ql Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSES/QL-LICENSE file.
-
-// Copyright 2015 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package stmts
-
-import (
-	"fmt"
-	"strings"
-
-	"github.com/juju/errors"
-	"github.com/pingcap/tidb/context"
-	"github.com/pingcap/tidb/mysql"
-	"github.com/pingcap/tidb/parser/coldef"
-	"github.com/pingcap/tidb/rset"
-	"github.com/pingcap/tidb/stmt"
-	"github.com/pingcap/tidb/util"
-	"github.com/pingcap/tidb/util/format"
-	"github.com/pingcap/tidb/util/sqlexec"
-)
-
-/************************************************************************************
- *  Account Management Statements
- *  https://dev.mysql.com/doc/refman/5.7/en/account-management-sql.html
- ************************************************************************************/
-var (
-	_ stmt.Statement = (*CreateUserStmt)(nil)
-	_ stmt.Statement = (*SetPwdStmt)(nil)
-)
-
-// CreateUserStmt creates user account.
-// See: https://dev.mysql.com/doc/refman/5.7/en/create-user.html
-type CreateUserStmt struct {
-	IfNotExists bool
-	Specs       []*coldef.UserSpecification
-
-	Text string
-}
-
-// Explain implements the stmt.Statement Explain interface.
-func (s *CreateUserStmt) Explain(ctx context.Context, w format.Formatter) {
-	w.Format("%s\n", s.Text)
-}
-
-// IsDDL implements the stmt.Statement IsDDL interface.
-func (s *CreateUserStmt) IsDDL() bool {
-	return true
-}
-
-// OriginText implements the stmt.Statement OriginText interface.
-func (s *CreateUserStmt) OriginText() string {
-	return s.Text
-}
-
-// SetText implements the stmt.Statement SetText interface.
-func (s *CreateUserStmt) SetText(text string) {
-	s.Text = text
-}
-
-func userExists(ctx context.Context, name string, host string) (bool, error) {
-	sql := fmt.Sprintf(`SELECT * FROM %s.%s WHERE User="%s" AND Host="%s";`, mysql.SystemDB, mysql.UserTable, name, host)
-	rs, err := ctx.(sqlexec.RestrictedSQLExecutor).ExecRestrictedSQL(ctx, sql)
-	if err != nil {
-		return false, errors.Trace(err)
-	}
-	defer rs.Close()
-	row, err := rs.Next()
-	if err != nil {
-		return false, errors.Trace(err)
-	}
-	return row != nil, nil
-}
-
-// parse user string into username and host
-// root@localhost -> roor, localhost
-func parseUser(user string) (string, string) {
-	strs := strings.Split(user, "@")
-	return strs[0], strs[1]
-}
-
-// Exec implements the stmt.Statement Exec interface.
-func (s *CreateUserStmt) Exec(ctx context.Context) (rset.Recordset, error) {
-	users := make([]string, 0, len(s.Specs))
-	for _, spec := range s.Specs {
-		userName, host := parseUser(spec.User)
-		exists, err1 := userExists(ctx, userName, host)
-		if err1 != nil {
-			return nil, errors.Trace(err1)
-		}
-		if exists {
-			if !s.IfNotExists {
-				return nil, errors.Errorf("Duplicate user")
-			}
-			continue
-		}
-		pwd := ""
-		if spec.AuthOpt.ByAuthString {
-			pwd = util.EncodePassword(spec.AuthOpt.AuthString)
-		} else {
-			pwd = util.EncodePassword(spec.AuthOpt.HashString)
-		}
-		user := fmt.Sprintf(`("%s", "%s", "%s")`, host, userName, pwd)
-		users = append(users, user)
-	}
-	if len(users) == 0 {
-		return nil, nil
-	}
-	sql := fmt.Sprintf(`INSERT INTO %s.%s (Host, User, Password) VALUES %s;`, mysql.SystemDB, mysql.UserTable, strings.Join(users, ", "))
-	_, err := ctx.(sqlexec.RestrictedSQLExecutor).ExecRestrictedSQL(ctx, sql)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	return nil, nil
-}
-
-// SetPwdStmt is a statement to assign a password to user account.
-// See: https://dev.mysql.com/doc/refman/5.7/en/set-password.html
-type SetPwdStmt struct {
-	User     string
-	Password string
-
-	Text string
-}
-
-// Explain implements the stmt.Statement Explain interface.
-func (s *SetPwdStmt) Explain(ctx context.Context, w format.Formatter) {
-	w.Format("%s\n", s.Text)
-}
-
-// IsDDL implements the stmt.Statement IsDDL interface.
-func (s *SetPwdStmt) IsDDL() bool {
-	return false
-}
-
-// OriginText implements the stmt.Statement OriginText interface.
-func (s *SetPwdStmt) OriginText() string {
-	return s.Text
-}
-
-// SetText implements the stmt.Statement SetText interface.
-func (s *SetPwdStmt) SetText(text string) {
-	s.Text = text
-}
-
-// Exec implements the stmt.Statement Exec interface.
-func (s *SetPwdStmt) Exec(ctx context.Context) (rset.Recordset, error) {
-	// TODO: If len(s.User) == 0, use CURRENT_USER()
-	userName, host := parseUser(s.User)
-	// Update mysql.user
-	sql := fmt.Sprintf(`UPDATE %s.%s SET password="%s" WHERE User="%s" AND Host="%s";`, mysql.SystemDB, mysql.UserTable, util.EncodePassword(s.Password), userName, host)
-	_, err := ctx.(sqlexec.RestrictedSQLExecutor).ExecRestrictedSQL(ctx, sql)
-	return nil, errors.Trace(err)
-}
